move some funnel utils out of static-viz
